LayoutTests/editing/selection/modify-up-on-rtl-wrapping-text.html

<!DOCTYPE html>
<html charset="utf-8">
<body>
<p>This tests clicking on the left of RTL text puts the caret at the end of the line.</p>
<pre id="console"></pre>
<script>

if (window.layoutTestController)
    layoutTestController.dumpAsText();

var tests = [
    {content: "&#1498; &#1500;&#1499;", width: "2.5ex", expected: [2, 4]},
    {content: "&#1490; &#1490;&#1499; &#1499;&#1506;&#1497;", width: "3.5ex", expected: [2, 5, 8]},
    {content: "&#1490;&#1499; &#1497;&#1490;&#1499; &#1497;&#1490;&#1499; &#1497;&#1490;&#1499; &#1497;&#1490;&#1499;", width: "5ex",
        expected: [3, 6, 10, 14, 18]},
];

function failed(message) {
    console.innerHTML += 'FAIL: ' + message + '\n';
}

function passed(message) {
    console.innerHTML += 'PASS: ' + message + '\n';
}

function runTest(container, test) {
    container.innerHTML = test.content;
    container.style.width = test.width;

    var lines = ['st', 'nd', 'rd', 'th'];
    window.getSelection().setPosition(container.firstChild, test.content.length);

    for (var i = 0; i < test.expected.length; i++) {
        if (!window.getSelection().isCollapsed)
            return failed('the selection was not collapsed');

        var range = window.getSelection().getRangeAt(0);
        if (range.startContainer != container.firstChild)
            return failed('caret was at a wrong container');

        var action = 'on ' + test.content + ', caret is at ' + range.startOffset;
        if (i == 0) action += ' initially';
        else if (i == 1) action += ' after moving upwards once';
        else if (i == 2) action += ' after moving upwards twice';
        else action += ' after moving upwards ' + i + ' times';
        if (range.startOffset != test.expected[test.expected.length - i - 1])
            return failed(action + ' but expected at ' + test.expected[i]);
        passed(action);
        window.getSelection().modify('move', 'backward', 'line')
    }
}

var console = document.getElementById('console');

var container = document.createElement('div');
container.contentEditable = true;
container.setAttribute('dir', 'rtl');
document.body.appendChild(container);

for (var i = 0; i < tests.length; i++)
    runTest(container, tests[i]);
//container.innerHTML = '';

</script>
</body>
</html>
