LayoutTests/editing/selection/modify-up-on-rtl-wrapping-text-expected.txt

This tests clicking on the left of RTL text puts the caret at the end of the line.

PASS: on ך לכ, caret is at 4 initially
PASS: on ך לכ, caret is at 2 after moving upwards once
PASS: on כ ככ כככ, caret is at 8 initially
PASS: on כ ככ כככ, caret is at 5 after moving upwards once
PASS: on כ ככ כככ, caret is at 2 after moving upwards twice
PASS: on גכ יגכ יגכ יגכ יגכ, caret is at 18 initially
PASS: on גכ יגכ יגכ יגכ יגכ, caret is at 14 after moving upwards once
PASS: on גכ יגכ יגכ יגכ יגכ, caret is at 10 after moving upwards twice
PASS: on גכ יגכ יגכ יגכ יגכ, caret is at 6 after moving upwards 3 times
PASS: on גכ יגכ יגכ יגכ יגכ, caret is at 3 after moving upwards 4 times
גכ יגכ יגכ יגכ יגכ
